refactored HBSS test cases
